+++ CHANGELOG.md
@@ -3,5 +3,10 @@ All notable changes to the "vscode-seeing-is-believing" extension will be docume
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## 0.0.2
+- Don't show error message twice
+- Refactor specs to test all commands against execution in non-ruby files
+- Rely on VS Code's error message if a command's returned promise is rejected
+
 ## 0.0.1
 - Initial release